extracted hotel from main directory to hotel folder
