+++ project/src/components/hotel/hotel-card.tsx
@@ -1,9 +1,9 @@
 import { Dispatch, SetStateAction } from 'react';
-import { Hotel } from '../../../types/hotel/hotel';
-import * as constants from '../../../constants';
-import { ApartmentType } from '../../../enums/apartment-type';
+import { Hotel } from '../../types/hotel/hotel';
+import * as constants from '../../constants';
+import { ApartmentType } from '../../enums/apartment-type';
 import { Link } from 'react-router-dom';
-import { getRoomUrl } from '../../../service/hotels';
+import { getRoomUrl } from '../../service/hotels';
 
 type HotelCardProps = {
   hotel: Hotel;
